+++ src/components/avatars/img-avatar/img-avatar.hbs
@@ -1,1 +1,2 @@
-<img src={{avatar}} alt={{name}} title={{name}} />
+<img src={{avatar}} alt={{name}} title={{name}} class="image-avatar" />
+<div class="profile__change-avatar">{{text}}</div>
