+++ src/components/buttons/button-submit/button-submit.scss
@@ -14,4 +14,5 @@
     color: #fff;
     font-weight: bolder;
     margin-bottom: 10px;
+    margin-top: 20px;
 }
