+++ src/components/chats/chat-main/chat-main-layout/chat-main-layout.hbs
@@ -3,4 +3,4 @@
     Выберите чат и начинайте общаться!
 </div>
 {{{messagefield}}}
-<div class="modal-place" id="modal-place"></div>
+<div class="modal-place hidden" id="modal-place"></div>
