+++ src/components/chats/chat-main/chat-main-layout/chat-main-layout.scss
@@ -391,4 +391,5 @@
     cursor: pointer;
     margin: 10px;
   }
+
 }
