+++ src/components/chats/chat-main/chat-main-menu/chat-main-menu.scss
@@ -1,0 +1,3 @@
+#open-menu li {
+    padding: 0;
+}
