+++ src/components/chats/chat-main/chat-main-menu/menu-items/menu-item.scss
@@ -4,3 +4,7 @@
     margin: 0;
     margin-bottom: 10px;
 }
+
+.chat-main__opened-menu #logout {
+    color: red;
+}
