+++ src/components/chats/chat-main/chat-main-modals/chat-main-modals.scss
@@ -3,7 +3,7 @@
     height: 100%;
     display: flex;
     flex-direction: column;
-    background-color: rgba(0, 0, 0, .2);
+    background-color: rgba(0, 0, 0, .4);
     position: absolute;
     z-index: 2050;
     top: 0;
@@ -15,8 +15,12 @@
     border: 1px solid gray;
     display: flex;
     flex-direction: column;
+    align-items: center;
+    justify-content: flex-start;
     padding: 10px;
     border-radius: 15px;
+    min-height: 200px;
+    min-width: 300px;
 }
 
 .result ul {
@@ -39,3 +43,20 @@
     background-image: url("./../../../../files/x_gray.png");
     cursor: pointer;
 }
+
+.modal h3 {
+    margin-bottom: 10px;
+}
+
+// .modal-place {
+//     width: 100vw;
+//     height: 100vh;
+//     background-color: rgba(0, 0, 0, .8);
+//     display: flex;
+//     align-items: center;
+//     justify-content: center;
+//     position: absolute;
+//     z-index: 1500;
+//     top: 0px;
+//     left: 0px;
+// }
