+++ src/components/chats/chat-main/chat-main-modals/create/index.ts
@@ -15,6 +15,7 @@ export function createChatModal() {
       name: "createChatSubmit",
       label: "Создать",
       type: "submit",
+      class: "submit-button small",
     }),
   });
