--- src/components/forms/form-changeavatar/modal-avatar/modal-avatar.hbs
@@ -1,20 +1,0 @@
-<div class="modal-wrapper center-center">
-  <div class="modal center-center" id="modalSearch">
-    <div class="close-button-wrapper">
-      <div class="close-button" id="close-button"></div>
-    </div>
-    <h3>Поменять аватарку чата</h3>
-    <form
-      id="changeChatAvatarForm"
-      method="post"
-      novalidate
-      enctype="multipart/form-data"
-    >
-      {{{avatar}}}
-      {{{input}}}
-      {{{button}}}
-    </form>
-    <div class="result" id="result">
-    </div>
-  </div>
-</div>
