--- src/components/forms/form-changeavatar/modal-avatar/modal-avatar.hbs.d.ts
@@ -1,3 +1,0 @@
-declare function modalAvatar (context: any, options?: any): string;
-
-export = modalAvatar;
