+++ src/components/forms/form-message/form-message.hbs
@@ -1,7 +1,7 @@
 <div class="chat-main__hr"></div>
 <form class="chat-main__message-form" novalidate>
-  <label for="message-field__file" title="Выберите файл"></label>
-  <input type="file" id="message-field__file" title="Выберите файл" />
+  {{!-- <label for="message-field__file" title="Выберите файл"></label>
+  <input type="file" id="message-field__file" title="Выберите файл" /> --}}
   {{{input}}}
   {{{button}}}
 </form>
