+++ src/components/forms/form-signin/form-signin.scss
@@ -15,6 +15,7 @@
 .form-signin input {
     width: 280px;
     margin: 0;
+    padding: 0;
     display: flex;
     flex-shrink: 0;
     box-sizing: content-box;
