+++ src/styles/_chat.scss
@@ -328,7 +328,7 @@
   }
 
   .chat-main__message-form input[type="text"] {
-    width: 600px;
+    width: 640px;
   }
 
   .chat-main__message-form input[type="file"],
