+++ src/styles/_forms.scss
@@ -3,7 +3,7 @@
     flex-direction: column;
     align-items: center;
     justify-content: space-between;
-    padding: 20px;
+    padding: 30px;
     width: 340px;
     box-sizing: border-box;
 }
